Hadamard streamer borrows instead of consuming inputs.
This was stupidly hard: the lifetimes of the batch of provers as output
by the EntryProduct argument have a lifetime that is passed from the
references of input streams onto the provers.
This implies that Vec<Box<dyn Prover + 'a>> can be extended only with
elements whose lifetime is at least 'a -- for instance,
elements defined after won't live long enough.