Update plookup relation
The plookup relation requires the left rotation of a polynomial,
which in turn requires the verifier to know the constant term
of the witness vector in order to simulate the evaluation responses.
Instead, we extend the witness for the entry product argument
with a leading zero, and then consider this shifted polynomial.
Now evaluations in x are simply xf(x) and shifted evaluations are f(x).